requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<5.0
psycopg2-binary>=2.9.9,<3.0
drf-spectacular>=0.27.0,<1.0
Pillow>=9.0.0
flake8>=4.0.0,<5.0
django-cors-headers>=4.0.0
django-allauth>=0.57.0
requests>=2.28.0
cryptography>=3.4.8
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
dj-database-url>=2.0.0
whitenoise>=6.5.0
cloudinary>=1.36.0
